Move codes to cli/bundler.rs
