+++ cli/Cargo.toml
@@ -63,7 +63,7 @@ serde_derive = "1.0.116"
 serde_json = { version = "1.0.57", features = [ "preserve_order" ] }
 sys-info = "0.7.0"
 sourcemap = "6.0.1"
-swc_bundler = "=0.7.2"
+swc_bundler = "=0.7.3"
 swc_common = { version = "=0.10.2", features = ["sourcemap"] }
 swc_ecmascript = { version = "=0.7.7", features = ["codegen", "dep_graph", "parser", "react", "transforms", "visit"] }
 tempfile = "3.1.0"
