+++ cli/main.rs
@@ -22,6 +22,7 @@ extern crate serde_derive;
 extern crate tokio;
 extern crate url;
 
+mod bundler;
 mod checksum;
 pub mod colors;
 mod coverage;
